refactor(pdf-visual): introduce QualityProfile dispatcher with fast profile
Split the captioner into a profile dispatcher plus a self-contained `fast`
profile (verbatim port of v0.14.0's SmolVLM-256M / IDEFICS3 implementation).
The dispatcher owns `env.cacheDir` and routes by `QualityProfile`; per-profile
modules stay free of the global side effect and can evolve independently.

The orchestrator now returns `{ pages, captions }` with `page.text` left
unmutated, and `prepareVisualPdfChunks` emits captions as dedicated chunks
wrapped with the historical `[Visual content on page N: ...]` marker. This
preserves caption integrity across the semantic chunker's sentence-boundary
splits.

Phase 1 of docs/plans/20260517-feature-vlm-visual-quality-mode.md:
  - `quality` profile placeholder throws fail-fast (wired in Phase 2)
  - `CaptionerConfig.modelName` (the legacy field on `ingest/visual.ts`'s
    interface) is currently passed through but ignored — `profile: 'fast'`
    is hardcoded at the captioner call site so default behavior matches
    v0.14.0. Phase 3 removes `modelName` from the caller chain.
