+++ src/pdf-visual/types.ts
@@ -29,17 +29,23 @@ export class VlmError extends Error {
 }
 
 /**
- * Captioner configuration. Model identifier and quantization variant are
- * pinned for v1; the only caller-tunable fields are the cache directory and
- * the optional execution device.
+ * Visual-quality profile selector. Each profile resolves to a self-contained
+ * captioner implementation under `captioners/`. `fast` ports the original
+ * SmolVLM-256M / IDEFICS3 captioner (lightweight default, ~250 MB cache);
+ * `quality` ports the Qwen2.5-VL-3B-Instruct-ONNX captioner (~2.9 GB cache,
+ * ~2× per-page inference, higher fidelity on figures with in-image text).
+ */
+export type QualityProfile = 'fast' | 'quality'
+
+/**
+ * Captioner configuration. The model identifier is no longer caller-tunable;
+ * it is resolved inside the selected `profile` so that prompt, chat template,
+ * processor signature, generation options, and model class stay coherent per
+ * profile.
  */
 export interface CaptionerConfig {
-  /**
-   * HuggingFace model identifier. Fixed by the caller in v1; the field is
-   * retained so a future model-family adapter can flip it from a literal to a
-   * resolver without changing this contract.
-   */
-  modelName: string
+  /** Visual-quality profile — selects the underlying VLM family. */
+  profile: QualityProfile
   /** Model cache directory (shared with the embedder via `env.cacheDir`). */
   cacheDir: string
   /** Execution device passed through to Transformers.js model loading. */
